+++ src/test/serialize_tests.cpp
@@ -84,18 +84,18 @@ BOOST_FIXTURE_TEST_SUITE(serialize_tests, BasicTestingSetup)
         BOOST_CHECK_EQUAL(sizeof(char), GetSerializeSize(bool(0), 0));
 
         // Sanity-check GetSerializeSize and c++ type matching
-        BOOST_CHECK_EQUAL(GetSerializeSize(char(0), 0), 1);
-        BOOST_CHECK_EQUAL(GetSerializeSize(int8_t(0), 0), 1);
-        BOOST_CHECK_EQUAL(GetSerializeSize(uint8_t(0), 0), 1);
-        BOOST_CHECK_EQUAL(GetSerializeSize(int16_t(0), 0), 2);
-        BOOST_CHECK_EQUAL(GetSerializeSize(uint16_t(0), 0), 2);
-        BOOST_CHECK_EQUAL(GetSerializeSize(int32_t(0), 0), 4);
-        BOOST_CHECK_EQUAL(GetSerializeSize(uint32_t(0), 0), 4);
-        BOOST_CHECK_EQUAL(GetSerializeSize(int64_t(0), 0), 8);
-        BOOST_CHECK_EQUAL(GetSerializeSize(uint64_t(0), 0), 8);
-        BOOST_CHECK_EQUAL(GetSerializeSize(float(0), 0), 4);
-        BOOST_CHECK_EQUAL(GetSerializeSize(double(0), 0), 8);
-        BOOST_CHECK_EQUAL(GetSerializeSize(bool(0), 0), 1);
+        BOOST_CHECK_EQUAL(GetSerializeSize(char(0), 0), (uint64_t)1);
+        BOOST_CHECK_EQUAL(GetSerializeSize(int8_t(0), 0), (uint64_t)1);
+        BOOST_CHECK_EQUAL(GetSerializeSize(uint8_t(0), 0), (uint64_t)1);
+        BOOST_CHECK_EQUAL(GetSerializeSize(int16_t(0), 0), (uint64_t)2);
+        BOOST_CHECK_EQUAL(GetSerializeSize(uint16_t(0), 0), (uint64_t)2);
+        BOOST_CHECK_EQUAL(GetSerializeSize(int32_t(0), 0), (uint64_t)4);
+        BOOST_CHECK_EQUAL(GetSerializeSize(uint32_t(0), 0), (uint64_t)4);
+        BOOST_CHECK_EQUAL(GetSerializeSize(int64_t(0), 0), (uint64_t)8);
+        BOOST_CHECK_EQUAL(GetSerializeSize(uint64_t(0), 0), (uint64_t)8);
+        BOOST_CHECK_EQUAL(GetSerializeSize(float(0), 0), (uint64_t)4);
+        BOOST_CHECK_EQUAL(GetSerializeSize(double(0), 0), (uint64_t)8);
+        BOOST_CHECK_EQUAL(GetSerializeSize(bool(0), 0), (uint64_t)1);
     }
 
     BOOST_AUTO_TEST_CASE(floats_conversion_test)
@@ -111,12 +111,12 @@ BOOST_FIXTURE_TEST_SUITE(serialize_tests, BasicTestingSetup)
         BOOST_CHECK_EQUAL(ser_uint32_to_float(0x40800000), 4.0F);
         BOOST_CHECK_EQUAL(ser_uint32_to_float(0x44444444), 785.066650390625F);
 
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(0.0F), 0x00000000);
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(0.5F), 0x3f000000);
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(1.0F), 0x3f800000);
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(2.0F), 0x40000000);
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(4.0F), 0x40800000);
-        BOOST_CHECK_EQUAL(ser_float_to_uint32(785.066650390625F), 0x44444444);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(0.0F), (uint64_t)0x00000000);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(0.5F), (uint64_t)0x3f000000);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(1.0F), (uint64_t)0x3f800000);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(2.0F), (uint64_t)0x40000000);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(4.0F), (uint64_t)0x40800000);
+        BOOST_CHECK_EQUAL(ser_float_to_uint32(785.066650390625F), (uint64_t)0x44444444);
     }
 
     BOOST_AUTO_TEST_CASE(doubles_conversion_test)
@@ -363,39 +363,39 @@ BOOST_FIXTURE_TEST_SUITE(serialize_tests, BasicTestingSetup)
 
         // Test inserting/deleting bytes.
         CDataStream ss(SER_DISK, 0);
-        BOOST_CHECK_EQUAL(ss.size(), 0);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)0);
 
         ss.write("\x00\x01\x02\xff", 4);
-        BOOST_CHECK_EQUAL(ss.size(), 4);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)4);
 
         char c = (char) 11;
 
         // Inserting at beginning/end/middle:
         ss.insert(ss.begin(), c);
-        BOOST_CHECK_EQUAL(ss.size(), 5);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)5);
         BOOST_CHECK_EQUAL(ss[0], c);
         BOOST_CHECK_EQUAL(ss[1], 0);
 
         ss.insert(ss.end(), c);
-        BOOST_CHECK_EQUAL(ss.size(), 6);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)6);
         BOOST_CHECK_EQUAL(ss[4], (char) 0xff);
         BOOST_CHECK_EQUAL(ss[5], c);
 
         ss.insert(ss.begin() + 2, c);
-        BOOST_CHECK_EQUAL(ss.size(), 7);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)7);
         BOOST_CHECK_EQUAL(ss[2], c);
 
         // Delete at beginning/end/middle
         ss.erase(ss.begin());
-        BOOST_CHECK_EQUAL(ss.size(), 6);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)6);
         BOOST_CHECK_EQUAL(ss[0], 0);
 
         ss.erase(ss.begin() + ss.size() - 1);
-        BOOST_CHECK_EQUAL(ss.size(), 5);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)5);
         BOOST_CHECK_EQUAL(ss[4], (char) 0xff);
 
         ss.erase(ss.begin() + 1);
-        BOOST_CHECK_EQUAL(ss.size(), 4);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)4);
         BOOST_CHECK_EQUAL(ss[0], 0);
         BOOST_CHECK_EQUAL(ss[1], 1);
         BOOST_CHECK_EQUAL(ss[2], 2);
@@ -404,7 +404,7 @@ BOOST_FIXTURE_TEST_SUITE(serialize_tests, BasicTestingSetup)
         // Make sure GetAndClear does the right thing:
         CSerializeData d;
         ss.GetAndClear(d);
-        BOOST_CHECK_EQUAL(ss.size(), 0);
+        BOOST_CHECK_EQUAL(ss.size(), (uint64_t)0);
     }
 
     BOOST_AUTO_TEST_CASE(class_methods_test)
