+++ src/wallet/test/wallet_tests.cpp
@@ -132,7 +132,7 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
             BOOST_CHECK_EQUAL(nValueRet, 37 * CENT);
             // and we can make 38 cents if we accept all new coins
             BOOST_CHECK(testWallet.SelectCoinsMinConf(38 * CENT, 1, 1, 0, vCoins, setCoinsRet, nValueRet));
-            BOOST_CHECK_EQUAL(nValueRet, 38 * CENT);
+            BOOST_CHECK_EQUAL(nValueRet, (int64_t)(38 * CENT));
 
             // try making 34 cents from 1,2,5,10,20 - we can't do it exactly
             BOOST_CHECK(testWallet.SelectCoinsMinConf(34 * CENT, 1, 1, 0, vCoins, setCoinsRet, nValueRet));
@@ -508,8 +508,8 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
             vpwallets[0] = &wallet;
             ::importwallet(request);
 
-            BOOST_CHECK_EQUAL(wallet.mapWallet.size(), 3L);
-            BOOST_CHECK_EQUAL(coinbaseTxns.size(), 103L);
+            BOOST_CHECK_EQUAL(wallet.mapWallet.size(), (uint64_t)3L);
+            BOOST_CHECK_EQUAL(coinbaseTxns.size(), (uint64_t)103L);
             for (size_t i = 0; i < coinbaseTxns.size(); ++i)
             {
                 bool found = wallet.GetWalletTx(coinbaseTxns[i].GetHash());
@@ -616,7 +616,7 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
         pwalletMain->AddDestData(dest, "rr1", "val_rr1");
 
         auto values = pwalletMain->GetDestValues("rr");
-        BOOST_CHECK_EQUAL(values.size(), 2L);
+        BOOST_CHECK_EQUAL(values.size(), (uint64_t)2L);
         BOOST_CHECK_EQUAL(values[0], "val_rr0");
         BOOST_CHECK_EQUAL(values[1], "val_rr1");
     }
@@ -675,9 +675,9 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
         // Confirm ListCoins initially returns 1 coin grouped under coinbaseKey
         // address.
         auto list = wallet->ListCoins();
-        BOOST_CHECK_EQUAL(list.size(), 1L);
+        BOOST_CHECK_EQUAL(list.size(), (uint64_t)1L);
         BOOST_CHECK_EQUAL(boost::get<CKeyID>(list.begin()->first).ToString(), coinbaseAddress);
-        BOOST_CHECK_EQUAL(list.begin()->second.size(), 1L);
+        BOOST_CHECK_EQUAL(list.begin()->second.size(), (uint64_t)1L);
 
         // Check initial balance from one mature coinbase transaction.
         BOOST_CHECK_EQUAL(5000 * COIN, wallet->GetAvailableBalance());
@@ -688,16 +688,16 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
         // pubkey.
         AddTx(CRecipient{GetScriptForRawPubKey({}), 1 * COIN, false /* subtract fee */});
         list = wallet->ListCoins();
-        BOOST_CHECK_EQUAL(list.size(), 1L);
+        BOOST_CHECK_EQUAL(list.size(), (uint64_t)1L);
         BOOST_CHECK_EQUAL(boost::get<CKeyID>(list.begin()->first).ToString(), coinbaseAddress);
         auto output = list.begin()->second[0].ToString();
         //std::cout << "OUTPUT:" << output << std::endl;
-        BOOST_CHECK_EQUAL(list.begin()->second.size(), 2L);
+        BOOST_CHECK_EQUAL(list.begin()->second.size(), (uint64_t)2L);
 
         // Lock both coins. Confirm number of available coins drops to 0.
         std::vector<COutput> available;
         wallet->AvailableCoins(available);
-        BOOST_CHECK_EQUAL(available.size(), 2L);
+        BOOST_CHECK_EQUAL(available.size(), (uint64_t)2L);
         for (const auto &group : list)
         {
             for (const auto &coin : group.second)
@@ -706,14 +706,14 @@ BOOST_FIXTURE_TEST_SUITE(wallet_tests, WalletTestingSetup)
             }
         }
         wallet->AvailableCoins(available);
-        BOOST_CHECK_EQUAL(available.size(), 0L);
+        BOOST_CHECK_EQUAL(available.size(), (uint64_t)0L);
 
         // Confirm ListCoins still returns same result as before, despite coins
         // being locked.
         list = wallet->ListCoins();
-        BOOST_CHECK_EQUAL(list.size(), 1L);
+        BOOST_CHECK_EQUAL(list.size(), (uint64_t)1L);
         BOOST_CHECK_EQUAL(boost::get<CKeyID>(list.begin()->first).ToString(), coinbaseAddress);
-        BOOST_CHECK_EQUAL(list.begin()->second.size(), 2L);
+        BOOST_CHECK_EQUAL(list.begin()->second.size(), (uint64_t)2L);
     }
 
 BOOST_AUTO_TEST_SUITE_END()
